back/requirements.txt
Flask>=3.0
Flask-Cors>=4.0
flask-restx>=1.3
numpy>=1.24
opencv-python-headless>=4.8
onnxruntime>=1.17
Pillow>=10.0
ultralytics>=8.4.0
